project1/cloud-run-template/requirements.txt
flask
gunicorn
google-cloud
google-cloud-core[grpc]
google-api-python-client
google-cloud-storage
google-cloud-runtimeconfig
google-cloud-pubsub
